Move tests to what they're testing. Naming conventions based on golang standards
